DTLZ Problems

MultiObjectiveProblems.DTLZ1 — Type

struct DTLZ1 <: DTLZ
    n_vars :: Int = 10;
    M :: Int = 6 ;
    k :: Int = n_vars + 1 - M;
end

Test problem defined in: Deb et. al - "Scalable Test Problems for Evolutionary Multiobjective Optimization".

Has n_vars decision variables and M objectives. The decision vector x is split into two parts: x[1:M-1] and x[M:end]. The latter has length k.

DTLZ1 has an affine-linear global Pareto-optimal front with Σfᵢ=0.5 and many local minima because a function g akin to Rastrigin's is used.

MultiObjectiveProblems.DTLZ2 — Type

struct DTLZ1 <: DTLZ
    n_vars :: Int = 15;
    M :: Int = 6 ;
    k :: Int = n_vars + 1 - M;
end

Test problem defined in: Deb et. al - "Scalable Test Problems for Evolutionary Multiobjective Optimization".

Has n_vars decision variables and M objectives. The decision vector x is split into two parts: x[1:M-1] and x[M:end]. The latter has length k.

DTLZ2 has spherical global Pareto-optimal front with Σfᵢ²=1.

MultiObjectiveProblems.DTLZ3 — Type

struct DTLZ1 <: DTLZ
    n_vars :: Int = 15;
    M :: Int = 6 ;
    k :: Int = n_vars + 1 - M;
end

Test problem defined in: Deb et. al - "Scalable Test Problems for Evolutionary Multiobjective Optimization".

Has n_vars decision variables and M objectives. The decision vector x is split into two parts: x[1:M-1] and x[M:end]. The latter has length k.

DTLZ3 has spherical global Pareto-optimal front with Σfᵢ²=1 and many local minima because a function g akin to Rastrigin's is used.

MultiObjectiveProblems.DTLZ4 — Type

struct DTLZ1 <: DTLZ
    n_vars :: Int = 15;
    M :: Int = 6 ;
    k :: Int = n_vars + 1 - M;
end

Test problem defined in: Deb et. al - "Scalable Test Problems for Evolutionary Multiobjective Optimization".

Has n_vars decision variables and M objectives. The decision vector x is split into two parts: x[1:M-1] and x[M:end]. The latter has length k.

DTLZ4 has spherical global Pareto-optimal front with Σfᵢ²=1 and solutions tend to lie near the $f_M-f₁$ plane.

MultiObjectiveProblems.DTLZ5 — Type

struct DTLZ1 <: DTLZ
    n_vars :: Int = 15;
    M :: Int = 6 ;
    k :: Int = n_vars + 1 - M;
end

Test problem defined in: Deb et. al - "Scalable Test Problems for Evolutionary Multiobjective Optimization".

Has n_vars decision variables and M objectives. The decision vector x is split into two parts: x[1:M-1] and x[M:end]. The latter has length k.

DTLZ5 has spherical Pareto-optimal front that is a curve.

MultiObjectiveProblems.DTLZ6 — Type

struct DTLZ1 <: DTLZ
    n_vars :: Int = 15;
    M :: Int = 6 ;
    k :: Int = n_vars + 1 - M;
end

Test problem defined in: Deb et. al - "Scalable Test Problems for Evolutionary Multiobjective Optimization".

Has n_vars decision variables and M objectives. The decision vector x is split into two parts: x[1:M-1] and x[M:end]. The latter has length k.

DTLZ6 has spherical Pareto-optimal front that is a curve and hard to reach.

MultiObjectiveProblems.DTLZ7 — Type

struct DTLZ1 <: DTLZ
    n_vars :: Int = 22;
    M :: Int = 3 ;
    k :: Int = n_vars + 1 - M;
end

Test problem defined in: Deb et. al - "Scalable Test Problems for Evolutionary Multiobjective Optimization".

Has n_vars decision variables and M objectives. The decision vector x is split into two parts: x[1:M-1] and x[M:end]. The latter has length k.

DTLZ7 has $2^{M-1}$ disconnected Pareto-optimal regions in search space. It is similar in construction to the ZDT problems.