ZDT Problems

MultiObjectiveProblems.ZDT — Type

abstract type ZDT <: MOP end

Abstract super type for the class of ZDT problems as defined in Zitzler et al. - 2000 - Comparison of Multiobjective Evolutionary Algorithms

These are biobjective problems with a variable number of decision variables n_vars >= 2. They share a common structure

\[\begin{aligned} f_1(\mathbf{x}) &= f( x_1 ),\\ f_2(\mathbf{x}) &= g( x_2, …, x_n ) · h( f_1( x_1 ), g( x_2, …, x_n ) ) \end{aligned}\]

MultiObjectiveProblems.ZDT1 — Type

ZDT1( n_vars :: Int )

Bi-Objective problem ZDT1 with convex Pareto Front as given in Zitzler et al. - 2000 - Comparison of Multiobjective Evolutionary Algorithms.

The objectives are given as

\[\begin{aligned} f_1( x_1 ) &= x_1,\\ f_2( x_1, …, x_n) &= g( x_2, …, x_n ) · h( f_1( x_1 ), g( x_2, …, x_n ) ), \\ g( x_2, …, x_n ) &= 1 + (9/(n-1)) · Σ_i x_i, \\ h( f1, g ) &= 1 - \sqrt{f_1/g}. \end{aligned}\]

MultiObjectiveProblems.ZDT2 — Type

ZDT2( n_vars :: Int )

Bi-Objective problem ZDT2 with nonconvex Pareto Front as given in Zitzler et al. - 2000 - Comparison of Multiobjective Evolutionary Algorithms.

The objectives are given as

\[\begin{aligned} f_1( x_1 ) &= x_1,\\ f_2( x_1, …, x_n) &= g( x_2, …, x_n ) · h( f_1( x_1 ), g( x_2, …, x_n ) ), \\ g( x_2, …, x_n ) &= 1 + (9/(n-1)) · Σ_i x_i, \\ h( f1, g ) &= 1 - (f_1/g)². \end{aligned}\]

MultiObjectiveProblems.ZDT3 — Type

ZDT3( n_vars :: Int )

Bi-Objective problem ZDT3 with discrete Pareto Front as given in Zitzler et al. - 2000 - Comparison of Multiobjective Evolutionary Algorithms.

The objectives are given as

\[\begin{aligned} f_1( x_1 ) &= x_1,\\ f_2( x_1, …, x_n) &= g( x_2, …, x_n ) · h( f_1( x_1 ), g( x_2, …, x_n ) ), \\ g( x_2, …, x_n ) &= 1 + (9/(n-1)) · Σ_i x_i, \\ h( f1, g ) &= 1 - \sqrt{f_1/g}- (f_1/g)·\sin( 10 π f_1 ) \end{aligned}\]

MultiObjectiveProblems.ZDT4 — Type

ZDT4( n_vars :: Int )

Bi-Objective problem ZDT4 with many local Pareto Fronts as given in Zitzler et al. - 2000 - Comparison of Multiobjective Evolutionary Algorithms.

The multimodality is due to $g$ being a version of Rastrigin's function.

The objectives are given as

\[\begin{aligned} f_1( x_1 ) &= x_1,\\ f_2( x_1, …, x_n) &= g( x_2, …, x_n ) · h( f_1( x_1 ), g( x_2, …, x_n ) ), \\ g( x_2, …, x_n ) &= 1 + (9/(n-1)) · Σ_i ( x_i^2 - \cos( 4 π x_i ) ), \\ h( f1, g ) &= 1 - \sqrt{f_1/g} \end{aligned}\]

MultiObjectiveProblems.ZDT6 — Type

ZDT6( n_vars :: Int )

Bi-Objective problem ZDT6 with biased Pareto Front as given in Zitzler et al. - 2000 - Comparison of Multiobjective Evolutionary Algorithms.

The objectives are given as

\[\begin{aligned} f_1( x_1 ) &= 1 - \exp(-4 x_1 ) \sin^6( 6 π x_1) ,\\ f_2( x_1, …, x_n) &= g( x_2, …, x_n ) · h( f_1( x_1 ), g( x_2, …, x_n ) ), \\ g( x_2, …, x_n ) &= 1 + 9 · Σ_i ( x_i / (n-1) )^{1/4}, \\ h( f1, g ) &= 1 - (f_1/g)² \end{aligned}\]

MultiObjectiveProblems.get_points — Function

get_points( samp_func, N; kwargs... )

Sample N points using samp_func::SamplingFunction. Additional kwargs are passed down to the problem-specfic implementations.

get_points( point_set, N; kwargs... )

Get all points using point_set::FixedPointSet.

get_points( sample_func, zdt, Val(:ParetoSet), N; method = :regular )

Return a list of points $x ∈ ℝⁿ$ where $x₁ = … = xₙ = 0$. Methods are :regular or :random.